+++ src/main/java/com/example/letalCosplay/repository/FacturaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.letalCosplay.model.Factura;
 
-public interface FacturaRepository extends JpaRepository<Factura, Integer> {
+public interface FacturaRepository extends JpaRepository<Factura, Long> {
 }
